coindecko/requirements.txt
numpy
pandas
requests
getpass
mysql-connector-python
